Add "A journey in software development" guest talk at UOIT on 2017-03-17.
A guest talk I gave at a second year CSE class at University of Ontario
Institute of Technology (UOIT) on March 17, 2017.

It's about Go (a quick introduction), and highlights of interesting
things I've done and seen while working on my master's degree,
at startups, on personal projects, and in open source.

A video of this talk was recorded and can be seen at
https://www.youtube.com/watch?v=XSd9OAHVXMI.