Add "A journey in software development" guest talk at UOIT on 2018-03-08.
A guest talk I gave at a second year CSE class at University of Ontario
Institute of Technology (UOIT) on March 8, 2018. This is a repeat of a
talk from previous year with minor updates, given to a new class of
students.

It's about Go (a quick introduction), and highlights of interesting
things I've done and seen while working on my master's degree,
at startups, on personal projects, and in open source.

A video of this talk was recorded and can be seen at
https://www.youtube.com/watch?v=XQqi6BwPhV8,
and the slides are publicly accessible at
https://dmitri.shuralyov.com/talks/2018/journey-in-software-development/journey.slide.